break down main into subfiles
