--- include/global.h
@@ -1,4 +1,0 @@
-#ifndef GLOBALS_HPP
-#define GLOBALS_HPP
-
-#endif
