+++ include/save.h
@@ -1,0 +1,9 @@
+#ifndef SAVE_H
+#define SAVE_H
+
+#include <string>
+
+void saveGame(const std::string& filename);
+void loadGame(const std::string& filename);
+
+#endif // SAVE_H
